pyproject.toml
[tool.poetry]
name = "trio-jsonrpc"
version = "0.1.0"
description = "JSON-RPC v2.0 for Trio"
authors = ["Mark E. Haase <mehaase@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sansio-jsonrpc = {path = "../sansio-jsonrpc"}
trio-websocket = "^0.8.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-trio = "^0.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
